Moved inner class FieldType
